Projeto atualizado

+++ projetos/Grupo B - Shoot Saber/ArduinoStarWars/Assets/Scripts/Player.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Player : MonoBehaviour
 {
-    public int life;
+    public float life;
     // Start is called before the first frame update
     void Start()
     {
